Move all modalButtons to common
