+++ SingularityUI/app/components/common/modalButtons/BounceModal.jsx
@@ -1,9 +1,9 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 
-import { BounceRequest } from '../../actions/api/requests';
+import { BounceRequest } from '../../../actions/api/requests';
 
-import FormModal from '../common/modal/FormModal';
+import FormModal from '../modal/FormModal';
 
 class BounceModal extends Component {
   static propTypes = {
