+++ SingularityUI/app/components/common/modalButtons/CancelDeployModal.jsx
