+++ SingularityUI/app/components/common/modalButtons/DisableHealthchecksButton.jsx
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from 'react';
 
 import { Glyphicon } from 'react-bootstrap';
 
-import { getClickComponent } from '../common/modal/ModalWrapper';
+import { getClickComponent } from '../modal/ModalWrapper';
 
 import DisableHealthchecksModal from './DisableHealthchecksModal';
