+++ SingularityUI/app/components/common/modalButtons/EnableHealthchecksButton.jsx
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from 'react';
 
 import { Glyphicon } from 'react-bootstrap';
 
-import { getClickComponent } from '../common/modal/ModalWrapper';
+import { getClickComponent } from '../modal/ModalWrapper';
 
 import EnableHealthchecksModal from './EnableHealthchecksModal';
