+++ SingularityUI/app/components/common/modalButtons/NewWebhookButton.jsx
@@ -4,7 +4,7 @@ import { Glyphicon } from 'react-bootstrap';
 import OverlayTrigger from 'react-bootstrap/lib/OverlayTrigger';
 import ToolTip from 'react-bootstrap/lib/Tooltip';
 
-import { getClickComponent } from '../common/modal/ModalWrapper';
+import { getClickComponent } from '../modal/ModalWrapper';
 
 import NewWebhookModal from './NewWebhookModal';
 
@@ -15,7 +15,7 @@ const newWebhookTooltip = (
 );
 
 const NewWebhookButton = ({children, user}) => {
-  const clickComponentData = {children};
+  const clickComponentData = {props: {children}}; // Tricks getClickComponent() into doing the right thing despite this being functional. Muahahaha
   return (
     <span>
       {getClickComponent(clickComponentData)}
