+++ SingularityUI/app/components/common/modalButtons/ScaleButton.jsx
@@ -4,30 +4,29 @@ import { Glyphicon } from 'react-bootstrap';
 import OverlayTrigger from 'react-bootstrap/lib/OverlayTrigger';
 import ToolTip from 'react-bootstrap/lib/Tooltip';
 
-import { getClickComponent } from '../common/modal/ModalWrapper';
+import { getClickComponent } from '../modal/ModalWrapper';
 
-import PauseModal from './PauseModal';
+import ScaleModal from './ScaleModal';
 
-const pauseTooltip = (
-  <ToolTip id="pause">
-    Pause
+const scaleTooltip = (
+  <ToolTip id="scale">
+    Scale
   </ToolTip>
 );
 
-export default class PauseButton extends Component {
-
+export default class ScaleButton extends Component {
   static propTypes = {
     requestId: PropTypes.string.isRequired,
-    isScheduled: PropTypes.bool.isRequired,
+    currentInstances: PropTypes.number,
     children: PropTypes.node,
     then: PropTypes.func
   };
 
   static defaultProps = {
     children: (
-      <OverlayTrigger placement="top" id="view-pause-overlay" overlay={pauseTooltip}>
-        <a>
-          <Glyphicon glyph="play" />
+      <OverlayTrigger placement="top" id="view-scale-overlay" overlay={scaleTooltip}>
+        <a title="Scale">
+          <Glyphicon glyph="signal" />
         </a>
       </OverlayTrigger>
     )
@@ -37,10 +36,10 @@ export default class PauseButton extends Component {
     return (
       <span>
         {getClickComponent(this)}
-        <PauseModal
+        <ScaleModal
           ref="modal"
           requestId={this.props.requestId}
-          isScheduled={this.props.isScheduled}
+          currentInstances={this.props.currentInstances}
           then={this.props.then}
         />
       </span>
